# apps/tenant/tenant-api/tenant-swagger.json
{"openapi":"3.0.0","paths":{"/api/config":{"get":{"operationId":"AppController_config","parameters":[],"responses":{"200":{"description":"Current BFF configuration data."}},"tags":["App Module"]}},"/api/health":{"get":{"operationId":"AppController_health","parameters":[],"responses":{"200":{"description":"Server is up."}},"tags":["App Module"]}},"/api/tenant":{"get":{"operationId":"TenantController_getTenants","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}}],"responses":{"200":{"description":"Get All Tenants","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantMapped"}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Tenant Module"],"security":[{"bearer":[]}]},"post":{"operationId":"TenantController_createTenant","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Tenant","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantDto"}}}},"responses":{"201":{"description":"Created Tenant"},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Tenant Module"],"security":[{"bearer":[]}]}},"/api/tenant/{id}":{"get":{"operationId":"TenantController_getTenantById","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get Tenant By Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Tenant Module"],"security":[{"bearer":[]}]},"put":{"operationId":"TenantController_updateTenant","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Tenant","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTenantDto"}}}},"responses":{"200":{"description":"Updated Tenant"},"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Tenant Module"],"security":[{"bearer":[]}]}},"/api/product-entitlement":{"get":{"operationId":"ProductEntitlementController_getProductEntitlements","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}}],"responses":{"200":{"description":"Get All Product Entitlements","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntitlementsList"}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Product Entitlement Module"],"security":[{"bearer":[]}]}},"/api/product-entitlement/tenants/{tenantId}":{"get":{"operationId":"ProductEntitlementController_getProductEntitlementsByTenantId","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get Product Entitlements By Id","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TenantEntitlement"}}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Product Entitlement Module"],"security":[{"bearer":[]}]}},"/api/product-entitlement/{productEntitlementId}/tenants/{tenantId}":{"post":{"operationId":"ProductEntitlementController_assignProductEntitlement","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"productEntitlementId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignTenantEntitlementDto"}}}},"responses":{"201":{"description":"Assign An Entitlement To A Tenant","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantEntitlement"}}}},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Product Entitlement Module"],"security":[{"bearer":[]}]}},"/api/product-entitlement/{productEntitlementId}/tenants/{tenantId}/activate":{"patch":{"operationId":"ProductEntitlementController_activateTenantEntitlement","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"productEntitlementId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Activates An Entitlement For A Tenant"},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Product Entitlement Module"],"security":[{"bearer":[]}]}},"/api/product-entitlement/{productEntitlementId}/tenants/{tenantId}/deactivate":{"patch":{"operationId":"ProductEntitlementController_deactivateTenantEntitlement","parameters":[{"name":"x-tenant-id","in":"header","description":"x-tenant-id","schema":{"type":"string"}},{"name":"productEntitlementId","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Deactivates An Entitlement For A Tenant"},"401":{"description":"Unauthorized"},"500":{"description":"Something has gone wrong.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantError"}}}}},"tags":["Product Entitlement Module"],"security":[{"bearer":[]}]}}},"info":{"title":"Tenant BFF","description":"Tenant back end for front end","version":"1.0","contact":{}},"tags":[{"name":"tenant","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"TenantError":{"type":"object","properties":{"message":{"type":"string"},"code":{"type":"string"},"status":{"type":"number"},"statusText":{"type":"string"},"data":{"type":"object"}},"required":["message","code","status","statusText","data"]},"TenantMapped":{"type":"object","properties":{"tenantId":{"type":"string"},"siteName":{"type":"string"},"createdDate":{"type":"string"},"tenantStatus":{"type":"string"}},"required":["tenantId","siteName","createdDate","tenantStatus"]},"Tenant":{"type":"object","properties":{"source_system":{"type":"string"},"tenantId":{"type":"string"},"siteName":{"type":"string"},"storageRegion":{"type":"string"},"tenantStatus":{"type":"string"},"tenantType":{"type":"string"},"ownerType":{"type":"string"},"cmpId":{"type":"string"},"customerName":{"type":"string"},"partnerName":{"type":"string"},"sourceSystem":{"type":"string"},"createdDate":{"type":"string"},"lastModifiedDate":{"type":"string"},"createdByUserId":{"type":"string"},"lastModifiedByUserId":{"type":"string"}},"required":["source_system","tenantId","siteName","storageRegion","tenantStatus","tenantType","ownerType","cmpId","customerName","partnerName","sourceSystem","createdDate","lastModifiedDate","createdByUserId","lastModifiedByUserId"]},"CreateTenantDto":{"type":"object","properties":{"site_name":{"type":"string"},"storage_region":{"type":"string"},"tenant_type":{"type":"string"},"owner_type":{"type":"string"},"cmp_id":{"type":"string"},"partner_name":{"type":"string"},"source_system":{"type":"string"}},"required":["site_name","storage_region","tenant_type","owner_type","cmp_id","partner_name","source_system"]},"UpdateTenantDto":{"type":"object","properties":{"site_name":{"type":"string"}}},"ProductEntitlementsList":{"type":"object","properties":{"itemsRequested":{"type":"number"},"itemsReturned":{"type":"number"},"itemsTotal":{"type":"number"},"offset":{"type":"number"},"items":{"type":"array","items":{"type":"array"}}},"required":["itemsRequested","itemsReturned","itemsTotal","offset","items"]},"TenantEntitlement":{"type":"object","properties":{"tenantId":{"type":"string"},"productEntitlementId":{"type":"string"},"productEntitlementName":{"type":"string"},"tenantEntitlementStatus":{"type":"string"},"assignmentDate":{"type":"string"},"amount":{"type":"number"},"assignmentSource":{"type":"string"},"sourceSystem":{"type":"string"},"createdDate":{"type":"string"},"lastModifiedDate":{"type":"string"},"createdByUserId":{"type":"string"},"lastModifiedByUserId":{"type":"string"}},"required":["tenantId","productEntitlementId","productEntitlementName","tenantEntitlementStatus","assignmentDate","amount","assignmentSource","sourceSystem","createdDate","lastModifiedDate","createdByUserId","lastModifiedByUserId"]},"AssignTenantEntitlementDto":{"type":"object","properties":{"assignment_date":{"type":"string"},"amount":{"type":"number"},"assignment_source":{"type":"string"},"source_system":{"type":"string"}},"required":["assignment_date","amount","assignment_source","source_system"]}}}}